+++ src/doc/template/index.html
@@ -70,7 +70,7 @@
 		</a>
 	</div>
 	<div class="common-relItem">
-		<a class="common-relItem_anchor" href="api/classes/rootapi.html" target="_blank">
+		<a class="common-relItem_anchor" href="api/classes/api_root.rootapi.html" target="_blank">
 			<div class="common-relItem_title">
 				API reference<span class="fas fa-external-link-square-alt"></span>
 			</div>
