+++ src/doc/template/misc.html
@@ -7,7 +7,7 @@
 	<div class="common-pageHeader_innerLayout">
 		<div class="common-pageHeader_titleLayout">
 			<h1 class="common-pageHeader_title">Misc</h1>
-			<p>Event handling, importing/exporting a preset, theming, and some tips.</p>
+			<p>Event handling, importing/exporting a preset, and some tips.</p>
 		</div>
 		<div class="common-pageHeader_paneLayout">
 			<div class="common-paneContainer common-paneContainer-misc"></div>
@@ -119,48 +119,6 @@ pane.addInput(PARAMS2, 'speed', {
 			<div class="common-paneContainer common-paneContainer-presetKey"></div>
 			<div class="common-demo_resultJsonLayout">
 				<pre class="common-demo_resultJson" data-presetKey></pre>
-			</div>
-		</div>
-	</div>
-</div>
-
-
-<h2 id="theming">Theming</h2>
-
-
-<p>Tweakpane now can be themed easily with CSS variable.</p>
-
-<div class="common-main_mediaLayout">
-	<div class="common-themes">
-		<div class="common-themes_themesLayout">
-			<div class="common-themes_themeLayout">
-				<div class="common-paneContainer common-paneContainer-defaultTheme"></div>
-			</div>
-			<div class="common-themes_themeLayout">
-				<div class="common-paneContainer common-paneContainer-lightTheme"></div>
-			</div>
-			<div class="common-themes_themeLayout">
-				<div class="common-paneContainer common-paneContainer-icebergTheme"></div>
-			</div>
-			<div class="common-themes_themeLayout">
-				<div class="common-paneContainer common-paneContainer-jetblackTheme"></div>
-			</div>
-		</div>
-		<div class="common-themes_mask common-themes_mask-start"></div>
-		<div class="common-themes_mask common-themes_mask-end"></div>
-	</div>
-</div>
-
-<p>Add <code>--tp-*</code> variables to the container element to change pane colors. All themeable variables are listed below:</p>
-
-<div class="common-main_mediaLayout">
-	<div class="common-demo">
-		<div class="common-demo_codeLayout">
-			<div class="common-codeBlock">
-				<pre><code class="html" data-themeCss></code></pre>
-				<div class="common-codeBlock_title">
-					.html
-				</div>
 			</div>
 		</div>
 	</div>
