+++ src/doc/template/partial/_global-nav.html
@@ -61,6 +61,9 @@
 			</ul>
 		</li>
 		<li class="common-menuItem">
+			<a class="common-menuItem_anchor" href="theming.html">Theming</a>
+		</li>
+		<li class="common-menuItem">
 			<a class="common-menuItem_anchor" href="misc.html">Misc</a>
 		</li>
 		<li class="common-menuItem">
@@ -72,15 +75,12 @@
 					<a class="common-submenuItem_anchor" href="misc.html#imex">Import/Export</a>
 				</li>
 				<li class="common-submenuItem">
-					<a class="common-submenuItem_anchor" href="misc.html#theming">Theming</a>
-				</li>
-				<li class="common-submenuItem">
 					<a class="common-submenuItem_anchor" href="misc.html#tips">Tips</a>
 				</li>
 			</ul>
 		</li>
 		<li class="common-menuItem">
-			<a class="common-menuItem_anchor" href="api/classes/rootapi.html" target="_blank">API reference<span class="fas fa-external-link-square-alt"></a>
+			<a class="common-menuItem_anchor" href="api/classes/api_root.rootapi.html" target="_blank">API reference<span class="fas fa-external-link-square-alt"></a>
 		</li>
 		<li class="common-menuItem">
 			<a class="common-menuItem_anchor" href="https://github.com/cocopon/tweakpane" target="_blank">GitHub<span class="fas fa-external-link-square-alt"></a>
