+++ src/doc/template/theming.html
@@ -1,0 +1,56 @@
+{% set pageId = 'theming' %}
+{% extends "_template.html" %}
+
+
+{% block pageHeader %}
+<div class="common-pageHeader">
+	<div class="common-pageHeader_innerLayout">
+		<div class="common-pageHeader_titleLayout">
+			<h1 class="common-pageHeader_title">Theming</h1>
+			<p>Tweakpane can be themed easily with CSS variable.</p>
+		</div>
+		<div class="common-pageHeader_paneLayout">
+			<div class="common-paneContainer common-paneContainer-header"></div>
+		</div>
+	</div>
+</div>
+{% endblock %}
+
+
+{% block content %}
+<div class="common-main_mediaLayout">
+	<div class="common-themes">
+		<div class="common-themes_themesLayout">
+			<div class="common-themes_themeLayout">
+				<div class="common-paneContainer common-paneContainer-defaultTheme"></div>
+			</div>
+			<div class="common-themes_themeLayout">
+				<div class="common-paneContainer common-paneContainer-lightTheme"></div>
+			</div>
+			<div class="common-themes_themeLayout">
+				<div class="common-paneContainer common-paneContainer-icebergTheme"></div>
+			</div>
+			<div class="common-themes_themeLayout">
+				<div class="common-paneContainer common-paneContainer-jetblackTheme"></div>
+			</div>
+		</div>
+		<div class="common-themes_mask common-themes_mask-start"></div>
+		<div class="common-themes_mask common-themes_mask-end"></div>
+	</div>
+</div>
+
+<p>Add <code>--tp-*</code> variables to the container element to change pane colors. All themeable variables are listed below:</p>
+
+<div class="common-main_mediaLayout">
+	<div class="common-demo">
+		<div class="common-demo_codeLayout">
+			<div class="common-codeBlock">
+				<pre><code class="html" data-themeCss></code></pre>
+				<div class="common-codeBlock_title">
+					.html
+				</div>
+			</div>
+		</div>
+	</div>
+</div>
+{% endblock %}
